web: refactor 5/7 node page charts
