+++ web/public/assets/js/app/value-helpers.js
@@ -1,0 +1,53 @@
+/*
+ * Copyright © 2025-26 l5yth & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Shared coercion helpers used across the node-page family of modules.
+ *
+ * Both ``node-page.js`` and ``node-page-charts.js`` historically defined a
+ * private ``numberOrNull`` and ``stringOrNull`` pair.  Lifting them into a
+ * shared module eliminates the duplication while keeping the helpers small
+ * and dependency-free for the chart submodules.
+ *
+ * @module value-helpers
+ */
+
+/**
+ * Attempt to coerce a raw value to a finite number.
+ *
+ * @param {*} value Raw value.
+ * @returns {number|null} Finite number or ``null``.
+ */
+export function numberOrNull(value) {
+  if (typeof value === 'number') {
+    return Number.isFinite(value) ? value : null;
+  }
+  if (value == null || value === '') return null;
+  const num = Number(value);
+  return Number.isFinite(num) ? num : null;
+}
+
+/**
+ * Coerce a raw value to a trimmed non-empty string.
+ *
+ * @param {*} value Raw value.
+ * @returns {string|null} Trimmed string or ``null``.
+ */
+export function stringOrNull(value) {
+  if (value == null) return null;
+  const str = String(value).trim();
+  return str.length === 0 ? null : str;
+}
